Move JsonDocumentBuilder into it’s own file
